refactor(config): split merge pipeline into focused modules
